feat(build): migrate 9 complex packages from bunup to @vertz/build (#2498)
- Migrate ui, ui-server, ui-auth, ui-primitives, theme-shadcn, docs, mdx, db, cli
- Auto-externalize devDependencies and optionalDependencies (matching bunup behavior)
- Set platform: 'node' for target: 'node' configs (ui-server bun-plugin)
- Update @vertz/ui and @vertz/ui-primitives exports from dist/src/ to dist/
- Fix native compiler fallback: replace Bun.Transpiler with esbuild transformSync
- Fix native compiler binary resolution: add import.meta.url fallback for Node ESM
