+++ packages/cli/package.json
@@ -28,7 +28,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "vtzx bunup && chmod +x dist/vertz.js",
+    "build": "vtzx vertz-build && chmod +x dist/vertz.js",
     "test": "vtz test",
     "test:watch": "vtz test --watch",
     "typecheck": "tsgo --noEmit -p tsconfig.typecheck.json"
@@ -52,7 +52,7 @@
     "@types/node": "^25.3.1",
     "@vertz/test": "workspace:*",
     "bun-types": "^1.3.10",
-    "bunup": "^0.16.23",
+    "@vertz/build": "workspace:^",
     "typescript": "^5.9.3"
   },
   "optionalDependencies": {
