+++ packages/db/package.json
@@ -49,7 +49,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "vtzx bunup",
+    "build": "vtzx vertz-build",
     "test": "vtz test",
     "test:watch": "vtz test --watch",
     "typecheck": "tsgo --noEmit -p tsconfig.typecheck.json"
@@ -71,7 +71,6 @@
     "@vertz/test": "workspace:*",
     "better-sqlite3": "^12.6.2",
     "bun-types": "^1.3.10",
-    "bunup": "^0.16.31",
     "postgres": "^3.4.8",
     "typescript": "^5.7.0"
   },
