+++ packages/docs/build.config.ts
@@ -1,9 +1,9 @@
+import { defineConfig } from '@vertz/build';
 import { createVertzLibraryPlugin } from '@vertz/ui-server';
-import { defineConfig } from 'bunup';
 
 export default defineConfig({
   entry: ['src/index.ts'],
-  dts: { inferTypes: true },
+  dts: true,
   plugins: [createVertzLibraryPlugin()],
   external: ['@vertz/ui', '@vertz/ui/internals'],
 });
