+++ packages/docs/package.json
@@ -26,7 +26,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "bunup",
+    "build": "vtzx vertz-build",
     "test": "vtz test",
     "typecheck": "tsgo --noEmit"
   },
@@ -49,7 +49,7 @@
     "@vertz/test": "workspace:*",
     "@vertz/ui-server": "workspace:^",
     "bun-types": "^1.3.10",
-    "bunup": "^0.16.31",
+    "@vertz/build": "workspace:^",
     "happy-dom": "^20.7.0",
     "typescript": "^5.7.0"
   },
