+++ packages/theme-shadcn/package.json
@@ -34,7 +34,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "bunup",
+    "build": "vtzx vertz-build",
     "test": "vtz test",
     "test:watch": "vtz test --watch",
     "typecheck": "tsgo --noEmit"
@@ -47,7 +47,7 @@
     "@happy-dom/global-registrator": "^20.7.0",
     "@vertz/test": "workspace:*",
     "@vertz/ui-server": "workspace:^",
-    "bunup": "^0.16.31",
+    "@vertz/build": "workspace:^",
     "happy-dom": "^20.7.0",
     "typescript": "^5.7.0"
   },
