+++ packages/ui-auth/build.config.ts
@@ -1,5 +1,5 @@
+import { defineConfig } from '@vertz/build';
 import { createVertzLibraryPlugin } from '@vertz/ui-server';
-import { defineConfig } from 'bunup';
 
 export default defineConfig({
   entry: ['src/index.ts'],
