+++ packages/ui-server/package.json
@@ -57,7 +57,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "vtzx bunup",
+    "build": "vtzx vertz-build",
     "test": "vtz test --timeout 60000 src/",
     "test:e2e": "vtzx playwright test",
     "typecheck": "tsgo --noEmit"
@@ -79,7 +79,7 @@
     "@vertz/codegen": "workspace:^",
     "@vertz/test": "workspace:*",
     "bun-types": "^1.3.10",
-    "bunup": "^0.16.31",
+    "@vertz/build": "workspace:^",
     "happy-dom": "^20.8.3",
     "typescript": "^5.7.0"
   },
