+++ packages/ui/package.json
@@ -16,52 +16,52 @@
   "sideEffects": [
     "dist/shared/*.js"
   ],
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/src/index.d.ts",
-      "import": "./dist/src/index.js"
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
     },
     "./internals": {
-      "types": "./dist/src/internals.d.ts",
-      "import": "./dist/src/internals.js"
+      "types": "./dist/internals.d.ts",
+      "import": "./dist/internals.js"
     },
     "./test": {
-      "types": "./dist/src/test/index.d.ts",
-      "import": "./dist/src/test/index.js"
+      "types": "./dist/test/index.d.ts",
+      "import": "./dist/test/index.js"
     },
     "./router": {
-      "types": "./dist/src/router/public.d.ts",
-      "import": "./dist/src/router/public.js"
+      "types": "./dist/router/public.d.ts",
+      "import": "./dist/router/public.js"
     },
     "./form": {
-      "types": "./dist/src/form/public.d.ts",
-      "import": "./dist/src/form/public.js"
+      "types": "./dist/form/public.d.ts",
+      "import": "./dist/form/public.js"
     },
     "./query": {
-      "types": "./dist/src/query/public.d.ts",
-      "import": "./dist/src/query/public.js"
+      "types": "./dist/query/public.d.ts",
+      "import": "./dist/query/public.js"
     },
     "./css": {
-      "types": "./dist/src/css/public.d.ts",
-      "import": "./dist/src/css/public.js"
+      "types": "./dist/css/public.d.ts",
+      "import": "./dist/css/public.js"
     },
     "./jsx-runtime": {
-      "types": "./dist/src/jsx-runtime/index.d.ts",
-      "import": "./dist/src/jsx-runtime/index.js"
+      "types": "./dist/jsx-runtime/index.d.ts",
+      "import": "./dist/jsx-runtime/index.js"
     },
     "./jsx-dev-runtime": {
-      "types": "./dist/src/jsx-runtime/index.d.ts",
-      "import": "./dist/src/jsx-runtime/index.js"
+      "types": "./dist/jsx-runtime/index.d.ts",
+      "import": "./dist/jsx-runtime/index.js"
     },
     "./auth": {
-      "types": "./dist/src/auth/public.d.ts",
-      "import": "./dist/src/auth/public.js"
+      "types": "./dist/auth/public.d.ts",
+      "import": "./dist/auth/public.js"
     },
     "./components": {
-      "types": "./dist/src/components/index.d.ts",
-      "import": "./dist/src/components/index.js"
+      "types": "./dist/components/index.d.ts",
+      "import": "./dist/components/index.js"
     },
     "./reactivity.json": "./reactivity.json"
   },
@@ -70,7 +70,7 @@
     "provenance": true
   },
   "scripts": {
-    "build": "vtzx bunup",
+    "build": "vtzx vertz-build",
     "test": "vtz test",
     "test:benchmark": "vtz test src/dom/__tests__/hydration-deferred-effects-bench.local.ts",
     "test:watch": "vtz test --watch",
@@ -83,7 +83,7 @@
     "@happy-dom/global-registrator": "^20.7.0",
     "@vertz/schema": "workspace:^",
     "@vertz/test": "workspace:*",
-    "bunup": "^0.16.31",
+    "@vertz/build": "workspace:^",
     "happy-dom": "^20.7.0",
     "typescript": "^5.7.0"
   },
